🔵 déplacement des types NAF & API vers reduce.algo2/populateNafAndApe.ts
