+++ dbmongo/js/globals.ts
@@ -2,26 +2,9 @@
 
 type ParPériode<T> = Record<Periode, T>
 
-type CodeAPE = string
-type CodeAPENiveau2 = string
-type CodeAPENiveau3 = string
-type CodeAPENiveau4 = string
-
-type CodeNAF = string
-
-type NAF = {
-  n2to1: Record<CodeAPENiveau2, CodeNAF>
-  n3to1: Record<CodeAPENiveau3, CodeNAF>
-  n4to1: Record<CodeAPENiveau4, CodeNAF>
-  n5to1: Record<CodeAPE, CodeNAF>
-  n1: Record<CodeNAF, string>
-  n2: Record<CodeAPENiveau2, string>
-  n3: Record<CodeAPENiveau3, string>
-  n4: Record<CodeAPENiveau4, string>
-  n5: Record<CodeAPE, string>
-}
-
 type Scope = "etablissement" | "entreprise"
+
+type CodeAPE = string
 
 // Détail des types de données
