+++ dbmongo/js/reduce.algo2/map.ts
@@ -11,7 +11,7 @@ import { defaillances } from "./defaillances"
 import { cotisationsdettes } from "./cotisationsdettes"
 import { ccsf } from "./ccsf"
 import { sirene } from "./sirene"
-import { populateNafAndApe } from "./populateNafAndApe"
+import { populateNafAndApe, NAF } from "./populateNafAndApe"
 import { cotisation } from "./cotisation"
 import { cibleApprentissage } from "./cibleApprentissage"
 import { entr_sirene, SortieSireneEntreprise } from "./entr_sirene"
